+++ helper/schema/grpc_provider.go
@@ -12,8 +12,8 @@ import (
 	ctyconvert "github.com/hashicorp/go-cty/cty/convert"
 	"github.com/hashicorp/go-cty/cty/msgpack"
 
-	"github.com/hashicorp/terraform-plugin-go/tfprotov5"
-	"github.com/hashicorp/terraform-plugin-go/tfprotov5/tftypes"
+	"github.com/hashicorp/terraform-plugin-go/tfprotov6"
+	"github.com/hashicorp/terraform-plugin-go/tfprotov6/tftypes"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/internal/configs/configschema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/internal/configs/hcl2shim"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/internal/plans/objchange"
@@ -61,30 +61,30 @@ func (s *GRPCProviderServer) StopContext(ctx context.Context) context.Context {
 	return stoppable
 }
 
-func (s *GRPCProviderServer) GetProviderSchema(_ context.Context, req *tfprotov5.GetProviderSchemaRequest) (*tfprotov5.GetProviderSchemaResponse, error) {
+func (s *GRPCProviderServer) GetProviderSchema(_ context.Context, req *tfprotov6.GetProviderSchemaRequest) (*tfprotov6.GetProviderSchemaResponse, error) {
 
-	resp := &tfprotov5.GetProviderSchemaResponse{
-		ResourceSchemas:   make(map[string]*tfprotov5.Schema),
-		DataSourceSchemas: make(map[string]*tfprotov5.Schema),
+	resp := &tfprotov6.GetProviderSchemaResponse{
+		ResourceSchemas:   make(map[string]*tfprotov6.Schema),
+		DataSourceSchemas: make(map[string]*tfprotov6.Schema),
 	}
 
-	resp.Provider = &tfprotov5.Schema{
+	resp.Provider = &tfprotov6.Schema{
 		Block: convert.ConfigSchemaToProto(s.getProviderSchemaBlock()),
 	}
 
-	resp.ProviderMeta = &tfprotov5.Schema{
+	resp.ProviderMeta = &tfprotov6.Schema{
 		Block: convert.ConfigSchemaToProto(s.getProviderMetaSchemaBlock()),
 	}
 
 	for typ, res := range s.provider.ResourcesMap {
-		resp.ResourceSchemas[typ] = &tfprotov5.Schema{
+		resp.ResourceSchemas[typ] = &tfprotov6.Schema{
 			Version: int64(res.SchemaVersion),
 			Block:   convert.ConfigSchemaToProto(res.CoreConfigSchema()),
 		}
 	}
 
 	for typ, dat := range s.provider.DataSourcesMap {
-		resp.DataSourceSchemas[typ] = &tfprotov5.Schema{
+		resp.DataSourceSchemas[typ] = &tfprotov6.Schema{
 			Version: int64(dat.SchemaVersion),
 			Block:   convert.ConfigSchemaToProto(dat.CoreConfigSchema()),
 		}
@@ -111,8 +111,8 @@ func (s *GRPCProviderServer) getDatasourceSchemaBlock(name string) *configschema
 	return dat.CoreConfigSchema()
 }
 
-func (s *GRPCProviderServer) PrepareProviderConfig(_ context.Context, req *tfprotov5.PrepareProviderConfigRequest) (*tfprotov5.PrepareProviderConfigResponse, error) {
-	resp := &tfprotov5.PrepareProviderConfigResponse{}
+func (s *GRPCProviderServer) ValidateProviderConfig(_ context.Context, req *tfprotov6.ValidateProviderConfigRequest) (*tfprotov6.ValidateProviderConfigResponse, error) {
+	resp := &tfprotov6.ValidateProviderConfigResponse{}
 
 	schemaBlock := s.getProviderSchemaBlock()
 
@@ -209,13 +209,13 @@ func (s *GRPCProviderServer) PrepareProviderConfig(_ context.Context, req *tfpro
 		return resp, nil
 	}
 
-	resp.PreparedConfig = &tfprotov5.DynamicValue{MsgPack: preparedConfigMP}
+	resp.PreparedConfig = &tfprotov6.DynamicValue{MsgPack: preparedConfigMP}
 
 	return resp, nil
 }
 
-func (s *GRPCProviderServer) ValidateResourceTypeConfig(_ context.Context, req *tfprotov5.ValidateResourceTypeConfigRequest) (*tfprotov5.ValidateResourceTypeConfigResponse, error) {
-	resp := &tfprotov5.ValidateResourceTypeConfigResponse{}
+func (s *GRPCProviderServer) ValidateResourceConfig(_ context.Context, req *tfprotov6.ValidateResourceConfigRequest) (*tfprotov6.ValidateResourceConfigResponse, error) {
+	resp := &tfprotov6.ValidateResourceConfigResponse{}
 
 	schemaBlock := s.getResourceSchemaBlock(req.TypeName)
 
@@ -232,8 +232,8 @@ func (s *GRPCProviderServer) ValidateResourceTypeConfig(_ context.Context, req *
 	return resp, nil
 }
 
-func (s *GRPCProviderServer) ValidateDataSourceConfig(_ context.Context, req *tfprotov5.ValidateDataSourceConfigRequest) (*tfprotov5.ValidateDataSourceConfigResponse, error) {
-	resp := &tfprotov5.ValidateDataSourceConfigResponse{}
+func (s *GRPCProviderServer) ValidateDataResourceConfig(_ context.Context, req *tfprotov6.ValidateDataResourceConfigRequest) (*tfprotov6.ValidateDataResourceConfigResponse, error) {
+	resp := &tfprotov6.ValidateDataResourceConfigResponse{}
 
 	schemaBlock := s.getDatasourceSchemaBlock(req.TypeName)
 
@@ -256,8 +256,8 @@ func (s *GRPCProviderServer) ValidateDataSourceConfig(_ context.Context, req *tf
 	return resp, nil
 }
 
-func (s *GRPCProviderServer) UpgradeResourceState(ctx context.Context, req *tfprotov5.UpgradeResourceStateRequest) (*tfprotov5.UpgradeResourceStateResponse, error) {
-	resp := &tfprotov5.UpgradeResourceStateResponse{}
+func (s *GRPCProviderServer) UpgradeResourceState(ctx context.Context, req *tfprotov6.UpgradeResourceStateRequest) (*tfprotov6.UpgradeResourceStateResponse, error) {
+	resp := &tfprotov6.UpgradeResourceStateResponse{}
 
 	res, ok := s.provider.ResourcesMap[req.TypeName]
 	if !ok {
@@ -332,7 +332,7 @@ func (s *GRPCProviderServer) UpgradeResourceState(ctx context.Context, req *tfpr
 		return resp, nil
 	}
 
-	resp.UpgradedState = &tfprotov5.DynamicValue{MsgPack: newStateMP}
+	resp.UpgradedState = &tfprotov6.DynamicValue{MsgPack: newStateMP}
 	return resp, nil
 }
 
@@ -482,7 +482,7 @@ func (s *GRPCProviderServer) removeAttributes(v interface{}, ty cty.Type) {
 	}
 }
 
-func (s *GRPCProviderServer) StopProvider(_ context.Context, _ *tfprotov5.StopProviderRequest) (*tfprotov5.StopProviderResponse, error) {
+func (s *GRPCProviderServer) StopProvider(_ context.Context, _ *tfprotov6.StopProviderRequest) (*tfprotov6.StopProviderResponse, error) {
 	s.stopMu.Lock()
 	defer s.stopMu.Unlock()
 
@@ -491,11 +491,11 @@ func (s *GRPCProviderServer) StopProvider(_ context.Context, _ *tfprotov5.StopPr
 	// reset the stop signal
 	s.stopCh = make(chan struct{})
 
-	return &tfprotov5.StopProviderResponse{}, nil
+	return &tfprotov6.StopProviderResponse{}, nil
 }
 
-func (s *GRPCProviderServer) ConfigureProvider(ctx context.Context, req *tfprotov5.ConfigureProviderRequest) (*tfprotov5.ConfigureProviderResponse, error) {
-	resp := &tfprotov5.ConfigureProviderResponse{}
+func (s *GRPCProviderServer) ConfigureProvider(ctx context.Context, req *tfprotov6.ConfigureProviderRequest) (*tfprotov6.ConfigureProviderResponse, error) {
+	resp := &tfprotov6.ConfigureProviderResponse{}
 
 	schemaBlock := s.getProviderSchemaBlock()
 
@@ -526,8 +526,8 @@ func (s *GRPCProviderServer) ConfigureProvider(ctx context.Context, req *tfproto
 	return resp, nil
 }
 
-func (s *GRPCProviderServer) ReadResource(ctx context.Context, req *tfprotov5.ReadResourceRequest) (*tfprotov5.ReadResourceResponse, error) {
-	resp := &tfprotov5.ReadResourceResponse{
+func (s *GRPCProviderServer) ReadResource(ctx context.Context, req *tfprotov6.ReadResourceRequest) (*tfprotov6.ReadResourceResponse, error) {
+	resp := &tfprotov6.ReadResourceResponse{
 		// helper/schema did previously handle private data during refresh, but
 		// core is now going to expect this to be maintained in order to
 		// persist it in the state.
@@ -586,7 +586,7 @@ func (s *GRPCProviderServer) ReadResource(ctx context.Context, req *tfprotov5.Re
 		if err != nil {
 			resp.Diagnostics = convert.AppendProtoDiag(resp.Diagnostics, err)
 		}
-		resp.NewState = &tfprotov5.DynamicValue{
+		resp.NewState = &tfprotov6.DynamicValue{
 			MsgPack: newStateMP,
 		}
 		return resp, nil
@@ -610,15 +610,15 @@ func (s *GRPCProviderServer) ReadResource(ctx context.Context, req *tfprotov5.Re
 		return resp, nil
 	}
 
-	resp.NewState = &tfprotov5.DynamicValue{
+	resp.NewState = &tfprotov6.DynamicValue{
 		MsgPack: newStateMP,
 	}
 
 	return resp, nil
 }
 
-func (s *GRPCProviderServer) PlanResourceChange(ctx context.Context, req *tfprotov5.PlanResourceChangeRequest) (*tfprotov5.PlanResourceChangeResponse, error) {
-	resp := &tfprotov5.PlanResourceChangeResponse{}
+func (s *GRPCProviderServer) PlanResourceChange(ctx context.Context, req *tfprotov6.PlanResourceChangeRequest) (*tfprotov6.PlanResourceChangeResponse, error) {
+	resp := &tfprotov6.PlanResourceChangeResponse{}
 
 	// This is a signal to Terraform Core that we're doing the best we can to
 	// shim the legacy type system of the SDK onto the Terraform type system
@@ -770,7 +770,7 @@ func (s *GRPCProviderServer) PlanResourceChange(ctx context.Context, req *tfprot
 		resp.Diagnostics = convert.AppendProtoDiag(resp.Diagnostics, err)
 		return resp, nil
 	}
-	resp.PlannedState = &tfprotov5.DynamicValue{
+	resp.PlannedState = &tfprotov6.DynamicValue{
 		MsgPack: plannedMP,
 	}
 
@@ -844,8 +844,8 @@ func (s *GRPCProviderServer) PlanResourceChange(ctx context.Context, req *tfprot
 	return resp, nil
 }
 
-func (s *GRPCProviderServer) ApplyResourceChange(ctx context.Context, req *tfprotov5.ApplyResourceChangeRequest) (*tfprotov5.ApplyResourceChangeResponse, error) {
-	resp := &tfprotov5.ApplyResourceChangeResponse{
+func (s *GRPCProviderServer) ApplyResourceChange(ctx context.Context, req *tfprotov6.ApplyResourceChangeRequest) (*tfprotov6.ApplyResourceChangeResponse, error) {
+	resp := &tfprotov6.ApplyResourceChangeResponse{
 		// Start with the existing state as a fallback
 		NewState: req.PriorState,
 	}
@@ -966,7 +966,7 @@ func (s *GRPCProviderServer) ApplyResourceChange(ctx context.Context, req *tfpro
 			resp.Diagnostics = convert.AppendProtoDiag(resp.Diagnostics, err)
 			return resp, nil
 		}
-		resp.NewState = &tfprotov5.DynamicValue{
+		resp.NewState = &tfprotov6.DynamicValue{
 			MsgPack: newStateMP,
 		}
 		return resp, nil
@@ -989,7 +989,7 @@ func (s *GRPCProviderServer) ApplyResourceChange(ctx context.Context, req *tfpro
 		resp.Diagnostics = convert.AppendProtoDiag(resp.Diagnostics, err)
 		return resp, nil
 	}
-	resp.NewState = &tfprotov5.DynamicValue{
+	resp.NewState = &tfprotov6.DynamicValue{
 		MsgPack: newStateMP,
 	}
 
@@ -1011,8 +1011,8 @@ func (s *GRPCProviderServer) ApplyResourceChange(ctx context.Context, req *tfpro
 	return resp, nil
 }
 
-func (s *GRPCProviderServer) ImportResourceState(ctx context.Context, req *tfprotov5.ImportResourceStateRequest) (*tfprotov5.ImportResourceStateResponse, error) {
-	resp := &tfprotov5.ImportResourceStateResponse{}
+func (s *GRPCProviderServer) ImportResourceState(ctx context.Context, req *tfprotov6.ImportResourceStateRequest) (*tfprotov6.ImportResourceStateResponse, error) {
+	resp := &tfprotov6.ImportResourceStateResponse{}
 
 	info := &terraform.InstanceInfo{
 		Type: req.TypeName,
@@ -1055,9 +1055,9 @@ func (s *GRPCProviderServer) ImportResourceState(ctx context.Context, req *tfpro
 			return resp, nil
 		}
 
-		importedResource := &tfprotov5.ImportedResource{
+		importedResource := &tfprotov6.ImportedResource{
 			TypeName: resourceType,
-			State: &tfprotov5.DynamicValue{
+			State: &tfprotov6.DynamicValue{
 				MsgPack: newStateMP,
 			},
 			Private: meta,
@@ -1069,8 +1069,8 @@ func (s *GRPCProviderServer) ImportResourceState(ctx context.Context, req *tfpro
 	return resp, nil
 }
 
-func (s *GRPCProviderServer) ReadDataSource(ctx context.Context, req *tfprotov5.ReadDataSourceRequest) (*tfprotov5.ReadDataSourceResponse, error) {
-	resp := &tfprotov5.ReadDataSourceResponse{}
+func (s *GRPCProviderServer) ReadDataSource(ctx context.Context, req *tfprotov6.ReadDataSourceRequest) (*tfprotov6.ReadDataSourceResponse, error) {
+	resp := &tfprotov6.ReadDataSourceResponse{}
 
 	schemaBlock := s.getDatasourceSchemaBlock(req.TypeName)
 
@@ -1121,7 +1121,7 @@ func (s *GRPCProviderServer) ReadDataSource(ctx context.Context, req *tfprotov5.
 		resp.Diagnostics = convert.AppendProtoDiag(resp.Diagnostics, err)
 		return resp, nil
 	}
-	resp.State = &tfprotov5.DynamicValue{
+	resp.State = &tfprotov6.DynamicValue{
 		MsgPack: newStateMP,
 	}
 	return resp, nil
@@ -1411,8 +1411,8 @@ func normalizeNullValues(dst, src cty.Value, apply bool) cty.Value {
 // appears in a list-like attribute (list, set, tuple) will present a nil value
 // to helper/schema which can panic. Return an error to the user in this case,
 // indicating the attribute with the null value.
-func validateConfigNulls(v cty.Value, path cty.Path) []*tfprotov5.Diagnostic {
-	var diags []*tfprotov5.Diagnostic
+func validateConfigNulls(v cty.Value, path cty.Path) []*tfprotov6.Diagnostic {
+	var diags []*tfprotov6.Diagnostic
 	if v.IsNull() || !v.IsKnown() {
 		return diags
 	}
@@ -1431,8 +1431,8 @@ func validateConfigNulls(v cty.Value, path cty.Path) []*tfprotov5.Diagnostic {
 					p = append(p, cty.IndexStep{Key: kv})
 				}
 
-				diags = append(diags, &tfprotov5.Diagnostic{
-					Severity:  tfprotov5.DiagnosticSeverityError,
+				diags = append(diags, &tfprotov6.Diagnostic{
+					Severity:  tfprotov6.DiagnosticSeverityError,
 					Summary:   "Null value found in list",
 					Detail:    "Null values are not allowed for this attribute value.",
 					Attribute: convert.PathToAttributePath(p),